+++ aws-api-appsync/src/main/java/com/amplifyframework/datastore/appsync/SerializedModel.java
@@ -151,7 +151,7 @@ public final class SerializedModel implements Model {
         return "SerializedModel{" +
             "id='" + modelId + '\'' +
             ", serializedData=" + serializedData +
-            ", modelSchema=" + modelSchema +
+            ", modelName=" + getModelName() +
             '}';
     }