+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## [Release 1.18.0](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.18.0)
+
+### Features
+- **aws-auth-cognito:** Allows userattributes in confirmSignIn (#1343)
+
+[See all changes between 1.17.8 and 1.18.0](https://github.com/aws-amplify/amplify-android/compare/release_v1.17.8...release_v1.18.0)
+
+## [Release 1.17.8](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.17.8)
+
+### Bug Fixes
+- **auth:** Add ConfirmSignUpOptions for confirmSignUp API method (#1357)
+- **storage:** remove duplicate error callback (#1366)
+
+[See all changes between 1.17.7 and 1.17.8](https://github.com/aws-amplify/amplify-android/compare/release_v1.17.7...release_v1.17.8)
+
+## [Release 1.17.7](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.17.7)
+
+### Bug Fixes
+- **api:** check for null ModelSchema to prevent crash in SerializedModel toString method (#1341)
+- **api:** default to mobile client for iam auth mode (#1351)
+- **Auth:** prevent multiple invocations of success callback for updateUserAttributes (#1339)
+
+### Miscellaneous
+- refactor:add enum to represent auth rule provider (#1320)
+
+[See all changes between 1.17.6 and 1.17.7](https://github.com/aws-amplify/amplify-android/compare/release_v1.17.6...release_v1.17.7)
+
+## [Release 1.17.6](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.17.6)
+
+### Bug Fixes
+- checkstyle failure on Windows (#1326)
+- **datastore:** save metadata when merging even if mutation outbox has pending item (#1319)
+- **datastore:** add syncExpression method to configuration builder that takes the modelName as a String (#1330)
+
+[See all changes between 1.17.5 and 1.17.6](https://github.com/aws-amplify/amplify-android/compare/release_v1.17.5...release_v1.17.6)
+
 ## [Release 1.17.5](https://github.com/aws-amplify/amplify-android/releases/tag/release_v1.17.5)
 
 ### Bug Fixes